pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = "test"
authors = ["解释 <xieshi@forchange.tech>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
pytest = "^5.4"
grpcio = "^1.30.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
